Allow additional values in `response_types`
The MCP spec requires clients and servers to use the authorization `code`
flow with PKCE, but servers may append other allowed flow types to the
`response_types` field. Previously, pydantic validation of the
`response_types` would raise an error if it contained anything other
than `code`.

This commit allows arbitrary strings in the `response_types` field, and instead
validates in the `RegistrationHandler` that the `response_types` field of a
registration request contains `code`.